Finished gameState, tested and should be ready to work Created asteroid bullet dummy file that in no way should be taken seriously

+++ bullet.h
@@ -1,13 +1,14 @@
-#include <SDL2/SDL.h>
-#include "object.cc"
 
 #ifndef BULLET_H
 #define BULLET_H
 
-class Bullet: public Object
+#include <SDL2/SDL.h>
+#include "object.h"
+
+class Bullet : public Object
 {
 public:
-	Bullet(int x_pos, int y_pos, double traj);
+	Bullet(double x_pos, double y_pos, double traj);
 	void drawSelf(SDL_Renderer* rend);
 };
